+++ blocks/edit/prose/index.js
@@ -24,7 +24,8 @@ import {
 
 // DA
 import prose2aem from '../../shared/prose2aem.js';
-import menu, { linkItem, getHeadingKeymap } from './plugins/menu.js';
+import menu, { getHeadingKeymap } from './plugins/menu/menu.js';
+import { linkItem } from './plugins/menu/linkItem.js';
 import imageDrop from './plugins/imageDrop.js';
 import linkConverter from './plugins/linkConverter.js';
 import sectionPasteHandler from './plugins/sectionPasteHandler.js';
